implement over-segment method

+++ src/icrengine.cpp
@@ -4,12 +4,13 @@
 #include <opencv2/imgproc/imgproc.hpp>
 
 #include <iosfwd>
+#include <string>
 #include <vector>
 
 using boost::filesystem::directory_iterator;
 using boost::filesystem::path;
 std::vector<DigitWidthStatistic> digitStatistics;
-
+std::string chqName;
 int main()
 {
 	computeDigitWidth("/media/thienlong/linux/CAR/cvl-digits/train", digitStatistics);
@@ -18,13 +19,24 @@ int main()
 		return 0;
 	}
 	icr::ICREngine engine;
+	int skip = 2;
+	int counter = 0;
 	for (directory_iterator iter(p), iterend; iter != iterend; ++iter) {
+		++counter;
+		if (counter <= skip) {
+			continue;
+		}
 		cv::Mat cheque = cv::imread(iter->path().string(), 0);
-//		cv::Mat cheque = cv::imread("/home/thienlong/cheque/500 Cheques/ValidChq/chq_00458_00.jpeg", 0);
+//		cv::Mat cheque = cv::imread("/home/thienlong/cheque/500 Cheques/ValidChq/chq_00234_00.jpeg", 0);
 		if (cheque.empty())
 			return -1;
-		std::cout << iter->path().filename().string() << ": ";
+		chqName = iter->path().filename().string();
+		std::size_t pos = chqName.find(".");
+
+		chqName = chqName.substr(0, pos);
+		std::cout << counter << "\t" << iter->path().filename().string() << ": ";
 		std::cout.flush();
+
 		std::cout << engine.recognite(cheque) << std::endl;
 		cv::waitKey(0);
 		cv::destroyAllWindows();
